Switches over to strtok.
Using strtok instead of scanf, because scanf swallows the newline and I was unable to find a solution for the empty command test. strtok is mostly implemented, I just need to implement some check that can remove the newline from the last token on the line. Slight changes made to the command struct to accomodate the new buffer that holds the tokens, as well as adding some booleans in for more readable control structures in main.